+++ ts/examples/chat/src/memory/knowproCommon.ts
@@ -1,14 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { ProgressBar } from "interactive-app";
+import { CommandMetadata, NamedArgs, ProgressBar } from "interactive-app";
 import * as knowLib from "knowledge-processor";
 import * as kp from "knowpro";
 import * as cm from "conversation-memory";
 import { MemoryConsoleWriter } from "../memoryWriter.js";
-import { addFileNameSuffixToPath } from "../common.js";
+import {
+    addFileNameSuffixToPath,
+    argToDate,
+    keyValuesFromNamedArgs,
+} from "../common.js";
 import path from "path";
-import { getFileName } from "typeagent";
+import { dateTime, getFileName } from "typeagent";
 import { TypeChatJsonTranslator } from "typechat";
 
 /**
@@ -203,4 +207,66 @@ export function* batchSemanticRefsByMessage(
     if (srs.length > 0) {
         yield [prevOrdinal, srs];
     }
+}
+
+export function createSearchGroup(
+    termArgs: string[],
+    namedArgs: NamedArgs,
+    commandDef: CommandMetadata,
+    op: "and" | "or" | "or_max",
+): kp.SearchTermGroup {
+    const searchTerms = kp.createSearchTerms(termArgs);
+    const propertyTerms = propertyTermsFromNamedArgs(namedArgs, commandDef);
+    return {
+        booleanOp: op,
+        terms: [...searchTerms, ...propertyTerms],
+    };
+}
+
+export function propertyTermsFromNamedArgs(
+    namedArgs: NamedArgs,
+    commandDef: CommandMetadata,
+): kp.PropertySearchTerm[] {
+    const keyValues = keyValuesFromNamedArgs(namedArgs, commandDef);
+    return kp.createPropertySearchTerms(keyValues);
+}
+
+export function whenFilterFromNamedArgs(
+    conversation: kp.IConversation,
+    namedArgs: NamedArgs,
+): kp.WhenFilter {
+    let filter: kp.WhenFilter = {
+        knowledgeType: namedArgs.ktype,
+    };
+    const dateRange = kp.getTimeRangeForConversation(conversation!);
+    if (dateRange) {
+        let startDate: Date | undefined;
+        let endDate: Date | undefined;
+        // Did they provide an explicit date range?
+        if (namedArgs.startDate || namedArgs.endDate) {
+            startDate = argToDate(namedArgs.startDate) ?? dateRange.start;
+            endDate = argToDate(namedArgs.endDate) ?? dateRange.end;
+        } else {
+            // They may have provided a relative date range
+            if (namedArgs.startMinute >= 0) {
+                startDate = dateTime.addMinutesToDate(
+                    dateRange.start,
+                    namedArgs.startMinute,
+                );
+            }
+            if (namedArgs.endMinute > 0) {
+                endDate = dateTime.addMinutesToDate(
+                    dateRange.start,
+                    namedArgs.endMinute,
+                );
+            }
+        }
+        if (startDate) {
+            filter.dateRange = {
+                start: startDate,
+                end: endDate,
+            };
+        }
+    }
+    return filter;
 }
