+++ ts/examples/chat/src/memory/knowproMemory.ts
@@ -18,18 +18,9 @@ import {
     runConsole,
     StopWatch,
 } from "interactive-app";
-import {
-    argToDate,
-    parseFreeAndNamedArguments,
-    createSearchGroup,
-} from "../common.js";
-import {
-    collections,
-    dateTime,
-    ensureDir,
-    isFilePath,
-    readJsonFile,
-} from "typeagent";
+import { parseFreeAndNamedArguments } from "../common.js";
+import { createSearchGroup, whenFilterFromNamedArgs } from "./knowproCommon.js";
+import { collections, ensureDir, isFilePath, readJsonFile } from "typeagent";
 import chalk from "chalk";
 import { KnowProPrinter } from "./knowproPrinter.js";
 import { createKnowproDataFrameCommands } from "./knowproDataFrame.js";
@@ -238,7 +229,7 @@ export async function createKnowproCommands(
                         ? "and"
                         : "or",
                 ),
-                when: whenFilterFromNamedArgs(namedArgs, commandDef),
+                when: whenFilterFromNamedArgs(conversation, namedArgs),
             };
             context.printer.writeSelectExpr(selectExpr);
             const timer = new StopWatch();
@@ -426,7 +417,7 @@ export async function createKnowproCommands(
                     ? "and"
                     : "or",
             ),
-            when: whenFilterFromNamedArgs(namedArgs, commandDef),
+            when: whenFilterFromNamedArgs(conversation, namedArgs),
         };
         const matches = await kp.searchConversation(
             conversation,
@@ -857,50 +848,6 @@ export async function createKnowproCommands(
         } else {
             context.printer.writeError(answerResult.message);
         }
-    }
-
-    function whenFilterFromNamedArgs(
-        namedArgs: NamedArgs,
-        commandDef: CommandMetadata,
-    ): kp.WhenFilter {
-        let filter: kp.WhenFilter = {
-            knowledgeType: namedArgs.ktype,
-        };
-        const conversation: kp.IConversation | undefined = context.conversation;
-        if (!conversation) {
-            throw new Error("No conversation loaded");
-        }
-        const dateRange = kp.getTimeRangeForConversation(conversation!);
-        if (dateRange) {
-            let startDate: Date | undefined;
-            let endDate: Date | undefined;
-            // Did they provide an explicit date range?
-            if (namedArgs.startDate || namedArgs.endDate) {
-                startDate = argToDate(namedArgs.startDate) ?? dateRange.start;
-                endDate = argToDate(namedArgs.endDate) ?? dateRange.end;
-            } else {
-                // They may have provided a relative date range
-                if (namedArgs.startMinute >= 0) {
-                    startDate = dateTime.addMinutesToDate(
-                        dateRange.start,
-                        namedArgs.startMinute,
-                    );
-                }
-                if (namedArgs.endMinute > 0) {
-                    endDate = dateTime.addMinutesToDate(
-                        dateRange.start,
-                        namedArgs.endMinute,
-                    );
-                }
-            }
-            if (startDate) {
-                filter.dateRange = {
-                    start: startDate,
-                    end: endDate,
-                };
-            }
-        }
-        return filter;
     }
 
     function ensureConversationLoaded(): kp.IConversation | undefined {
